Make structured functions properly check device/dtype of explicit out args
Somehow I forgot to add these checks.  Now they're in here.  Thanks
ngimel for noticing.

This is probably a slight efficiency hit on TensorIterator, which is
probably already doing all these checks.  Would be good to follow up
on this, though it may not be easily fixable with the TI rewrite.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: a5abeeae8700c764b43227f3f17cd70b756f492d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/55150