[RFC] Unify device check and guard in kernel wrapper
Pull Request resolved: https://github.com/pytorch/pytorch/pull/55241

For kernels run on CUDA backend, the following preparations are required:

(1) Make sure all the tensors are on the same device (except for
0-dimension tensor, which is allowed to stay in CPU). We will denote
this device as `common_device`.

(2) Acquire the device guard on `common_device`.

Today, these two preparations are done seperately. (1) is done in kernels
in an adhoc way, while (2) is done in the kernel wrapper (e.g. `RegisterCUDA.cpp`).

The kernel wrapper assumes (1) will be done in kernel, so it just pick up
the first tensor to acquire device gurad. There are at least two issues with
the current appraoch:

1. Kernel may not impelemnt the device check, or implement in inconsistent way.
In fact, there are a few implementations in codebase to do the checks.

2. If the first tensor happnened to be a 0-dimension CPU tensor, (2) will acquire
the wrong device guard. We have seen issues before
(db2b273d1fe00d63bc8932e6136fea83efef1de4) thus kernel might need to re-acquire
device guard in such situation, which is fragile..
ghstack-source-id: 125666279

Differential Revision: [D27540071](https://our.internmc.facebook.com/intern/diff/D27540071/)